requirements.txt
streamlit
python-docx
gTTS

